refactor(waf): centralize waf rule construction
