+++ lib/migrate.js
@@ -29,11 +29,8 @@
  */
 Object.defineProperty(exports, "__esModule", { value: true });
 const fs = require('fs');
-const path = require('path');
 const yaml = require('js-yaml');
-function resolveAbsolute(inputPath, cwd) {
-    return path.isAbsolute(inputPath) ? inputPath : path.join(cwd, inputPath);
-}
+const { resolveAbsolute } = require('../emitter');
 function migratePolicy(opts = {}) {
     opts = opts || {};
     const cwd = opts.cwd || process.cwd();
